move participantSort into StageProps
