+++ src/components/StageProps.ts
@@ -1,4 +1,4 @@
-import { Room } from "livekit-client";
+import { Participant, Room } from "livekit-client";
 import { RoomState } from "../useRoom";
 import { ControlsProps } from "./ControlsView";
 import { ParticipantProps } from "./ParticipantView";
@@ -8,4 +8,5 @@ export interface StageProps {
   participantRenderer?: (props: ParticipantProps) => React.ReactElement | null;
   controlRenderer?: (props: ControlsProps) => React.ReactElement | null;
   onLeave?: (room: Room) => void;
+  sortParticipants?: (participants: Participant[]) => void;
 }
