+++ src/components/StageView.tsx
@@ -13,6 +13,7 @@ import styles from "./styles.module.css";
 
 export const StageView = (stageProps: StageProps) => {
   const isMobile = useMediaQuery({ query: "(max-width: 800px)" });
+
   const { room, participants } = stageProps.roomState;
   const context = useContext(DisplayContext);
